fix: restore board entry point
