+++ src/ui/board.ts
@@ -1,12 +1,9 @@
-// mainBoard.ts — merged & de-conflicted
-
 import * as Server from '@/server';
 import {
   DB,
   KS,
   STATE,
   loadStaff,
-  saveStaff,
   CURRENT_SCHEMA_VERSION,
   migrateActiveBoard,
   setActiveBoardCache,
@@ -14,32 +11,33 @@ import {
   mergeBoards,
   type Staff,
   type ActiveBoard,
-  type DraftShift,
   getConfig,
-  saveConfig,
   type Config,
 } from '@/state';
 import { notifyUpdate, onUpdate } from '@/state/sync';
-
-import { setNurseCache, labelFromId } from '@/utils/names';
-import { renderWeather } from './widgets';
-import { renderPhysicians, renderPhysicianPopup } from './physicians';
-import { nurseTile } from './nurseTile';
-import { createDebouncer } from '@/utils/debouncedSave';
-import './mainBoard/boardLayout.css';
-import {
-  startBreak,
-  endBreak,
-  moveSlot,
-  upsertSlot,
-  removeSlot,
-  type Slot,
-} from '@/slots';
+import { setNurseCache } from '@/utils/names';
+import { renderWeather } from '@/ui/widgets';
+import { renderPhysicians, renderPhysicianPopup } from '@/ui/physicians';
+import '@/ui/mainBoard/boardLayout.css';
 import { normalizeActiveZones, type ZoneDef } from '@/utils/zones';
-import { showBanner, showToast } from '@/ui/banner';
-import { openAssignDialog } from '@/ui/assignDialog';
+import { showToast } from '@/ui/banner';
 
-const RECENT_MS = 15 * 60 * 1000;
+import { createPatientCareTeamPanel, renderLeadership } from '@/ui/board/patientCareTeam';
+import { createAssignmentsPanel, renderAssignments, defaultEnd } from '@/ui/board/assignments';
+import { createCommentsPanel, wireComments } from '@/ui/board/comments';
+import { createWeatherPanel } from '@/ui/board/weather';
+import {
+  createIncomingPanel,
+  renderIncoming,
+  autoAssignArrivals,
+} from '@/ui/board/incoming';
+import { createOffgoingPanel, renderOffgoing } from '@/ui/board/offgoing';
+import { createPhysiciansPanel } from '@/ui/board/physicians';
+import type { Slot } from '@/slots';
+
+export { defaultEnd };
+export { renderLeadership } from '@/ui/board/patientCareTeam';
+
 const toMin = (hhmm: string): number => {
   const [h, m] = hhmm.split(':').map(Number);
   return h * 60 + m;
@@ -49,29 +47,6 @@ const toHHMM = (min: number): string => {
   const m = min % 60;
   return `${String(h).padStart(2, '0')}:${String(m).padStart(2, '0')}`;
 };
-export const defaultEnd = (start: string): string =>
-  toHHMM((toMin(start) + 12 * 60) % 1440);
-let clockHandler: (() => void) | null = null;
-
-const offlineQueue: ActiveBoard[] = [];
-
-async function flushQueuedSaves(): Promise<void> {
-  while (offlineQueue.length) {
-    const board = offlineQueue[0];
-    try {
-      await Server.save('active', board);
-      offlineQueue.shift();
-    } catch {
-      break;
-    }
-  }
-}
-
-window.addEventListener('online', () => {
-  void flushQueuedSaves();
-});
-
-// --- helpers ---------------------------------------------------------------
 
 function buildEmptyActive(
   dateISO: string,
@@ -94,7 +69,19 @@ function buildEmptyActive(
   };
 }
 
-// --- top-level render ------------------------------------------------------
+let clockHandler: (() => void) | null = null;
+const offlineQueue: ActiveBoard[] = [];
+async function flushQueuedSaves(): Promise<void> {
+  while (offlineQueue.length) {
+    const board = offlineQueue[0];
+    try {
+      await Server.save('active', board);
+      offlineQueue.shift();
+    } catch {
+      break;
+    }
+  }
+}
 
 /** Render the main board view. */
 export async function renderBoard(
@@ -108,7 +95,6 @@ export async function renderBoard(
     const staff: Staff[] = await loadStaff();
     setNurseCache(staff);
 
-    // Load or initialize active shift tuple
     const saveKey = KS.ACTIVE(ctx.dateISO, ctx.shift);
 
     let active: ActiveBoard | undefined = await DB.get<ActiveBoard>(saveKey);
@@ -146,63 +132,30 @@ export async function renderBoard(
       showToast('Using local data; changes may not persist');
     }
 
-    // Layout
-    root.innerHTML = `
-      <div class="layout" data-testid="main-board">
-        <div class="col col-left">
-          <section class="panel">
-            <h3>Patient Care Team</h3>
-            <div class="slots lead">
-              <div id="slot-charge"></div>
-              <div id="slot-triage"></div>
-              <div id="slot-secretary"></div>
-            </div>
-            <div id="pct-zones" class="zones-grid"></div>
-          </section>
+    // Layout assembly
+    root.innerHTML = '';
+    const layout = document.createElement('div');
+    layout.className = 'layout';
+    layout.setAttribute('data-testid', 'main-board');
 
-          <section class="panel">
-            <h3>Assignments</h3>
-            <div id="zones" class="zones-grid"></div>
-          </section>
+    const left = document.createElement('div');
+    left.className = 'col col-left';
+    left.appendChild(createPatientCareTeamPanel());
+    left.appendChild(createAssignmentsPanel());
+    left.appendChild(createCommentsPanel());
+    layout.appendChild(left);
 
-          <section class="panel">
-            <h3>Comments</h3>
-            <textarea
-              id="comments"
-              class="input"
-              placeholder="Current Status..."
-              data-gramm="false"
-            ></textarea>
-          </section>
-        </div>
+    const right = document.createElement('div');
+    right.className = 'col col-right';
+    right.appendChild(createWeatherPanel());
+    right.appendChild(createIncomingPanel());
+    right.appendChild(createOffgoingPanel());
+    right.appendChild(createPhysiciansPanel());
+    layout.appendChild(right);
 
-        <div class="col col-right">
-          <section id="weather" class="panel">
-            <h3>Weather</h3>
-            <div id="weather-body"></div>
-          </section>
+    root.appendChild(layout);
 
-          <section id="incoming-panel" class="panel">
-            <h3>Incoming (click to toggle arrived)</h3>
-            <button id="add-incoming" class="btn">+ Add</button>
-            <div id="incoming" class="min-h-40"></div>
-          </section>
-
-          <section id="offgoing-panel" class="panel">
-            <h3>Offgoing (kept 40 min)</h3>
-            <div id="offgoing"></div>
-          </section>
-
-          <section class="panel">
-            <h3>Physicians</h3>
-            <div id="phys"></div>
-            <button id="phys-next7" class="btn">Next 7 days</button>
-          </section>
-        </div>
-      </div>
-    `;
-
-    // Save immediately to IndexedDB/broadcast, debounce server writes
+    // Save logic
     let saveTimer: ReturnType<typeof setTimeout> | undefined;
 
     const saveLocal = () => {
@@ -230,34 +183,33 @@ export async function renderBoard(
 
     const refresh = () => {
       renderLeadership(active!, staff, queueSave, root, refresh);
-      renderZones(active!, cfg, staff, queueSave, root);
+      renderAssignments(active!, cfg, staff, queueSave, root);
     };
 
     refresh();
     wireComments(active!, queueSave);
-      await renderIncoming(active!, staff, queueSave);
-  renderOffgoing(active!, queueSave);
+    await renderIncoming(active!, staff, queueSave);
+    renderOffgoing(active!, queueSave);
 
-      const checkArrivals = () => {
-        if (autoAssignArrivals(active!, cfg)) {
-          queueSave();
-          void renderIncoming(active!, staff, queueSave);
-          renderZones(active!, cfg, staff, queueSave, root);
-        }
-      };
-      if (clockHandler) document.removeEventListener('clock-tick', clockHandler);
-      clockHandler = checkArrivals;
-      document.addEventListener('clock-tick', checkArrivals);
-      checkArrivals();
+    const checkArrivals = () => {
+      if (autoAssignArrivals(active!, cfg)) {
+        queueSave();
+        void renderIncoming(active!, staff, queueSave);
+        renderAssignments(active!, cfg, staff, queueSave, root);
+      }
+    };
+    if (clockHandler) document.removeEventListener('clock-tick', clockHandler);
+    clockHandler = checkArrivals;
+    document.addEventListener('clock-tick', checkArrivals);
+    checkArrivals();
 
-      const weatherBody = document.getElementById('weather-body');
-      if (weatherBody) await renderWeather(weatherBody);
+    const weatherBody = document.getElementById('weather-body');
+    if (weatherBody) await renderWeather(weatherBody);
 
     await renderPhysicians(
       document.getElementById('phys') as HTMLElement,
       ctx.dateISO
     );
-
     const btn = document.getElementById('phys-next7');
     btn?.addEventListener('click', () => {
       renderPhysicianPopup(ctx.dateISO, 7);
@@ -284,13 +236,12 @@ export async function renderBoard(
       refresh();
     });
 
-    // Re-render on config changes (e.g., zone list or colors)
     document.addEventListener('config-changed', () => {
       const c = getConfig();
       normalizeActiveZones(active!, c.zones);
       queueSave();
       renderLeadership(active!, staff, queueSave, root, refresh);
-      renderZones(active!, c, staff, queueSave, root);
+      renderAssignments(active!, c, staff, queueSave, root);
     });
   } catch (err) {
     console.error(err);
@@ -306,578 +257,3 @@ export async function renderBoard(
     });
   }
 }
-
-// --- leadership ------------------------------------------------------------
-
-export function renderLeadership(
-  active: ActiveBoard,
-  staff: Staff[],
-  save: () => void,
-  root: ParentNode,
-  rerender: () => void
-): void {
-  const cfg = getConfig();
-  const chargeEl = root.querySelector('#slot-charge') as HTMLElement | null;
-  const triageEl = root.querySelector('#slot-triage') as HTMLElement | null;
-  const secEl = root.querySelector('#slot-secretary') as HTMLElement | null;
-  if (!chargeEl || !triageEl || !secEl) {
-    console.warn('Missing leadership slot element');
-    return;
-  }
-
-  chargeEl.textContent = labelFromId(active.charge?.nurseId);
-  triageEl.textContent = labelFromId(active.triage?.nurseId);
-  secEl.textContent = labelFromId(active.admin?.nurseId);
-
-  chargeEl.style.display =
-    active.charge?.nurseId || cfg.showPinned?.charge ? '' : 'none';
-  triageEl.style.display =
-    active.triage?.nurseId || cfg.showPinned?.triage ? '' : 'none';
-  secEl.style.display = active.admin?.nurseId ? '' : 'none';
-
-  chargeEl.onclick = () =>
-    assignLeadDialog(active, staff, save, 'charge', rerender);
-  triageEl.onclick = () =>
-    assignLeadDialog(active, staff, save, 'triage', rerender);
-  secEl.onclick = () =>
-    assignLeadDialog(active, staff, save, 'admin', rerender);
-}
-
-function assignLeadDialog(
-  board: ActiveBoard,
-  staffList: Staff[],
-  save: () => void,
-  role: 'charge' | 'triage' | 'admin',
-  rerender: () => void
-): void {
-  const overlay = document.createElement('div');
-  overlay.className = 'manage-overlay';
-  const roleLabel =
-    role === 'charge' ? 'Charge Nurse' : role === 'triage' ? 'Triage Nurse' : 'Unit Secretary';
-  overlay.innerHTML = `
-    <div class="manage-dialog">
-      <h3>Assign ${roleLabel}</h3>
-      <select id="lead-select">
-        <option value="">Unassigned</option>
-        ${staffList
-          .map((s) => `<option value="${s.id}">${s.name || s.id}</option>`)
-          .join('')}
-      </select>
-      <div class="dialog-actions">
-        <button id="lead-save" class="btn">Save</button>
-        <button id="lead-cancel" class="btn">Cancel</button>
-      </div>
-    </div>`;
-  document.body.appendChild(overlay);
-
-  const sel = overlay.querySelector('#lead-select') as HTMLSelectElement;
-  const currentId =
-    role === 'charge'
-      ? board.charge?.nurseId
-      : role === 'triage'
-      ? board.triage?.nurseId
-      : board.admin?.nurseId;
-  if (currentId) sel.value = currentId;
-
-  overlay.querySelector('#lead-cancel')!.addEventListener('click', () =>
-    overlay.remove()
-  );
-  overlay.querySelector('#lead-save')!.addEventListener('click', () => {
-    const id = sel.value;
-    if (id) {
-      const moved = upsertSlot(board, role, { nurseId: id });
-      if (moved) showBanner('Previous assignment cleared');
-    } else {
-      if (removeSlot(board, role)) showBanner('Assignment cleared');
-    }
-    save();
-    overlay.remove();
-    rerender();
-  });
-}
-
-// --- zones & tiles ---------------------------------------------------------
-
-function renderZones(
-  active: ActiveBoard,
-  cfg: Config,
-  staff: Staff[],
-  save: () => void,
-  root: ParentNode
-): void {
-  const pctCont = root.querySelector('#pct-zones') as HTMLElement | null;
-  const cont = root.querySelector('#zones') as HTMLElement | null;
-  if (!pctCont || !cont) {
-    console.warn('Missing zones container');
-    return;
-  }
-  pctCont.innerHTML = '';
-  cont.innerHTML = '';
-  pctCont.style.minHeight = '40px';
-  cont.style.minHeight = '40px';
-  const zones: ZoneDef[] = cfg.zones || [];
-
-  zones.forEach((z: ZoneDef, i: number) => {
-    const zName = z.name;
-    const section = document.createElement('section');
-    section.className = 'zone-card';
-    section.setAttribute('data-testid', 'zone-card');
-    section.draggable = true;
-    section.addEventListener('dragstart', (e) => {
-      const ev = e as DragEvent;
-      ev.dataTransfer?.setData('zone-index', String(i));
-    });
-
-    section.addEventListener('dragover', (e) => {
-      e.preventDefault();
-      e.stopPropagation();
-    });
-
-    section.addEventListener('drop', async (e) => {
-      e.preventDefault();
-      e.stopPropagation();
-      const dt = (e as DragEvent).dataTransfer;
-      const nurseId = dt?.getData('incoming-id');
-      if (nurseId) {
-        const rawEnd = prompt('Shift end time (HH:MM)?')?.trim();
-        const end = rawEnd && /^\d{4}$/.test(rawEnd)
-          ? rawEnd.replace(/(\d{2})(\d{2})/, '$1:$2')
-          : rawEnd;
-        const slot: Slot = {
-          nurseId,
-          startHHMM: STATE.clockHHMM,
-          endTimeOverrideHHMM: end || defaultEnd(STATE.clockHHMM),
-        };
-        const moved = upsertSlot(active, { zone: z.name }, slot);
-        if (moved) showBanner('Previous assignment cleared');
-        active.incoming = active.incoming.filter((i) => i.nurseId !== nurseId);
-        save();
-        renderIncoming(active, staff, save);
-        renderZones(active, cfg, staff, save, root);
-        return;
-      }
-
-      const zoneIdxStr = dt?.getData('zone-index');
-      if (!zoneIdxStr) return;
-      const fromIdx = Number(zoneIdxStr);
-      if (isNaN(fromIdx) || fromIdx === i) return;
-      const [moved] = cfg.zones.splice(fromIdx, 1);
-      moved.pct = z.pct;
-      cfg.zones.splice(i, 0, moved);
-      await saveConfig({ zones: cfg.zones });
-      renderZones(active, cfg, staff, save, root);
-    });
-
-    // Highlight color: explicit or from palette
-    const explicit = z.color || cfg.zoneColors?.[zName];
-    if (explicit) {
-      section.style.setProperty('--zone-color', explicit);
-    } else {
-      const zi = (i % 8) + 1;
-      section.style.setProperty('--zone-color', `var(--zone-bg-${zi})`);
-    }
-
-    const title = document.createElement('h2');
-    title.className = 'zone-card__title';
-    title.textContent = zName;
-    section.appendChild(title);
-
-    const editBtn = document.createElement('button');
-    editBtn.textContent = '⚙';
-    editBtn.className = 'zone-card__edit';
-    editBtn.title = 'Edit zone';
-    editBtn.addEventListener('click', async () => {
-      const val = prompt('Rename zone', z.name)?.trim();
-      if (val && val !== z.name) {
-        if (cfg.zones.some((zz) => zz.name === val)) {
-          alert('A zone with that name already exists.');
-          return;
-        }
-        const idx = cfg.zones.findIndex((zz) => zz.name === z.name);
-        cfg.zones[idx].name = val;
-        if (cfg.zoneColors && cfg.zoneColors[z.name]) {
-          cfg.zoneColors[val] = cfg.zoneColors[z.name];
-          delete cfg.zoneColors[z.name];
-        }
-        active.zones[val] = active.zones[z.name] || [];
-        delete active.zones[z.name];
-        await saveConfig({ zones: cfg.zones, zoneColors: cfg.zoneColors });
-        document.dispatchEvent(new Event('config-changed'));
-        await save();
-        renderZones(active, cfg, staff, save, root);
-      }
-    });
-    section.appendChild(editBtn);
-
-    const body = document.createElement('div');
-    body.className = 'zone-card__body';
-
-    (active.zones[zName] || []).forEach((s: Slot, idx: number) => {
-      let st = staff.find((n) => n.id === s.nurseId);
-      if (!st) {
-        console.warn('Unknown staffId', s.nurseId);
-        st = {
-          id: s.nurseId,
-          name: s.nurseId,
-          role: 'nurse',
-          type: 'other',
-        } as Staff;
-      }
-
-      const row = document.createElement('div');
-      row.className = 'nurse-row';
-
-      const tileWrapper = document.createElement('div');
-      tileWrapper.innerHTML = nurseTile(s, {
-        id: st.id,
-        name: st.name,
-        role: st.role || 'nurse',
-        type: st.type || 'other',
-      } as Staff);
-      const tile = tileWrapper.firstElementChild as HTMLElement;
-      row.appendChild(tile);
-      if (s.assignedTs && Date.now() - s.assignedTs < RECENT_MS) {
-        tile.classList.add('recent-assignment');
-        const remaining = RECENT_MS - (Date.now() - s.assignedTs);
-        setTimeout(() => tile.classList.remove('recent-assignment'), remaining);
-      }
-
-      const btn = document.createElement('button');
-      btn.textContent = 'Manage';
-      btn.className = 'btn';
-      btn.addEventListener('click', () =>
-        manageSlot(
-          s,
-          st,
-          staff,
-          save,
-          () => renderZones(active, cfg, staff, save, root),
-          z.name,
-          idx,
-          active,
-          cfg
-        )
-      );
-
-      row.appendChild(btn);
-      body.appendChild(row);
-    });
-
-    const hasSlots = (active.zones[zName] || []).length > 0;
-    const addBtn = document.createElement('button');
-    addBtn.textContent = '+';
-    addBtn.className = hasSlots
-      ? 'zone-card__add zone-card__add--small'
-      : 'zone-card__add zone-card__add--large';
-    addBtn.title = 'Add staff';
-    addBtn.addEventListener('click', () => {
-      openAssignDialog(staff, (id) => {
-        const slot: Slot = {
-          nurseId: id,
-          startHHMM: STATE.clockHHMM,
-          endTimeOverrideHHMM: defaultEnd(STATE.clockHHMM),
-        };
-        const moved = upsertSlot(active, { zone: z.name }, slot);
-        if (moved) showBanner('Previous assignment cleared');
-        save();
-        renderZones(active, cfg, staff, save, root);
-      });
-    });
-    if (hasSlots) {
-      section.appendChild(addBtn);
-    } else {
-      body.appendChild(addBtn);
-    }
-
-    section.appendChild(body);
-    (z.pct ? pctCont : cont).appendChild(section);
-  });
-
-  const enableDrop = (container: HTMLElement, pct: boolean) => {
-    container.addEventListener('dragover', (e) => e.preventDefault());
-    container.addEventListener('drop', async (e) => {
-      e.preventDefault();
-      const zoneIdxStr = (e as DragEvent).dataTransfer?.getData('zone-index');
-      if (!zoneIdxStr) return;
-      const idx = Number(zoneIdxStr);
-      if (isNaN(idx)) return;
-      cfg.zones[idx].pct = pct;
-      await saveConfig({ zones: cfg.zones });
-      renderZones(active, cfg, staff, save, root);
-    });
-  };
-
-  enableDrop(pctCont, true);
-  enableDrop(cont, false);
-}
-
-// --- comments --------------------------------------------------------------
-
-function wireComments(active: ActiveBoard, save: () => void) {
-  const el = document.getElementById('comments') as HTMLTextAreaElement | null;
-  if (!el) return;
-
-  el.value = active.comments || '';
-  el.disabled = STATE.locked;
-  const commit = () => {
-    active.comments = el.value;
-    save();
-  };
-
-  const debounced = createDebouncer(
-    () => {
-      active.comments = el.value;
-    },
-    () => save()
-  );
-
-  el.addEventListener('input', debounced);
-  el.addEventListener('blur', commit);
-}
-
-// --- incoming & offgoing ---------------------------------------------------
-
-async function renderIncoming(
-  active: ActiveBoard,
-  staffList: Staff[],
-  save: () => void
-) {
-  const cont = document.getElementById('incoming')!;
-  cont.innerHTML = '';
-
-  // auto-populate from draft schedule 40 min before start time
-  const draft = await DB.get<DraftShift>(KS.DRAFT(STATE.dateISO, STATE.shift));
-  let changed = false;
-  if (draft) {
-    const slots: Slot[] = [];
-    if (draft.charge) slots.push(draft.charge);
-    if (draft.triage) slots.push(draft.triage);
-    if (draft.admin) slots.push(draft.admin);
-    for (const arr of Object.values(draft.zones)) slots.push(...arr);
-    const now = toMin(STATE.clockHHMM);
-    for (const s of slots) {
-      if (!s.startHHMM) continue;
-      const diff = toMin(s.startHHMM) - now;
-      if (diff <= 40 && diff >= 0) {
-        if (!active.incoming.some((i) => i.nurseId === s.nurseId)) {
-          active.incoming.push({ nurseId: s.nurseId, eta: s.startHHMM });
-          changed = true;
-        }
-      }
-    }
-  }
-  if (changed) save();
-
-  if (active.incoming.length === 0) {
-    cont.innerHTML = '<div class="incoming-placeholder"></div>';
-  } else {
-    active.incoming.forEach((inc) => {
-      const st =
-        staffList.find((s) => s.id === inc.nurseId) ||
-        ({ id: inc.nurseId, name: inc.nurseId, role: 'nurse', type: 'other' } as Staff);
-
-      const row = document.createElement('div');
-      row.className = 'nurse-row';
-
-      const tileWrapper = document.createElement('div');
-      tileWrapper.innerHTML = nurseTile({ nurseId: inc.nurseId }, st);
-      const card = tileWrapper.firstElementChild as HTMLElement;
-      card.draggable = true;
-      card.addEventListener('dragstart', (e) => {
-        (e as DragEvent).dataTransfer?.setData('incoming-id', inc.nurseId);
-      });
-      const toggleArrived = () => {
-        if (STATE.locked) return;
-        inc.arrived = !inc.arrived;
-        const cfg = getConfig();
-        const moved = autoAssignArrivals(active, cfg);
-        save();
-        void renderIncoming(active, staffList, save);
-        if (moved) {
-          renderZones(
-            active,
-            cfg,
-            staffList,
-            save,
-            document.getElementById('panel')!
-          );
-        }
-      };
-      card.addEventListener('click', toggleArrived);
-      row.appendChild(card);
-
-      const eta = document.createElement('div');
-      eta.textContent = `${inc.eta}${inc.arrived ? ' ✓' : ''}`;
-      eta.addEventListener('click', toggleArrived);
-      row.appendChild(eta);
-
-      cont.appendChild(row);
-    });
-  }
-
-  const btn = document.getElementById('add-incoming') as HTMLButtonElement | null;
-  if (btn) {
-    btn.disabled = STATE.locked;
-    btn.onclick = () => {
-      if (STATE.locked) return;
-      openAssignDialog(staffList, (id) => {
-        const overlay = document.createElement('div');
-        overlay.className = 'manage-overlay';
-        overlay.innerHTML = `
-          <div class="manage-dialog">
-            <h3>Add Incoming</h3>
-            <p>${labelFromId(id)}</p>
-            <label>ETA <input id="inc-eta" placeholder="HH:MM"></label>
-            <div class="dialog-actions">
-              <button id="inc-save" class="btn">Save</button>
-              <button id="inc-cancel" class="btn">Cancel</button>
-            </div>
-          </div>`;
-        document.body.appendChild(overlay);
-        const etaEl = overlay.querySelector('#inc-eta') as HTMLInputElement | null;
-        const saveBtn = overlay.querySelector('#inc-save') as HTMLButtonElement | null;
-        const cancelBtn = overlay.querySelector('#inc-cancel') as HTMLButtonElement | null;
-        saveBtn?.addEventListener('click', () => {
-          active.incoming.push({ nurseId: id, eta: (etaEl?.value || '').trim() });
-          save();
-          void renderIncoming(active, staffList, save);
-          overlay.remove();
-        });
-        cancelBtn?.addEventListener('click', () => overlay.remove());
-      });
-    };
-  }
-}
-
-function renderOffgoing(active: ActiveBoard, save: () => void) {
-  const cont = document.getElementById('offgoing')!;
-  const cutoff = Date.now() - 60 * 60 * 1000; // 60 min
-  active.offgoing = (active.offgoing || []).filter((o) => o.ts > cutoff);
-
-  cont.innerHTML = '';
-  for (const o of active.offgoing) {
-    const div = document.createElement('div');
-    div.textContent = labelFromId(o.nurseId);
-    cont.appendChild(div);
-  }
-  save();
-}
-
-function autoAssignArrivals(active: ActiveBoard, cfg: Config): boolean {
-  const auxName = 'Aux 1';
-  if (!cfg.zones?.some((z) => z.name === auxName) || !active.zones[auxName]) {
-    return false;
-  }
-  const now = toMin(STATE.clockHHMM);
-  let moved = false;
-  active.incoming = active.incoming.filter((inc) => {
-    if (inc.arrived && inc.eta && toMin(inc.eta) <= now) {
-      upsertSlot(active, {
-        zone: auxName,
-      }, {
-        nurseId: inc.nurseId,
-        startHHMM: inc.eta,
-        endTimeOverrideHHMM: defaultEnd(inc.eta),
-      });
-      moved = true;
-      return false;
-    }
-    return true;
-  });
-  return moved;
-}
-
-// --- manage overlay --------------------------------------------------------
-
-function manageSlot(
-  slot: Slot,
-  st: Staff | undefined,
-  staffList: Staff[],
-  save: () => void,
-  rerender: () => void,
-  zone: string,
-  index: number,
-  board: ActiveBoard,
-  cfg: Config
-): void {
-  if (!st) return;
-
-  const startValue = slot.startHHMM || STATE.clockHHMM;
-  const pad2 = (n: number) => n.toString().padStart(2, '0');
-  const duration = cfg.shiftDurations?.[STATE.shift] ?? 12;
-  let endValue = slot.endTimeOverrideHHMM && /^\d{4}$/.test(slot.endTimeOverrideHHMM)
-    ? slot.endTimeOverrideHHMM.replace(/(\d{2})(\d{2})/, '$1:$2')
-    : slot.endTimeOverrideHHMM || '';
-  if (!endValue && startValue) {
-    const [sh, sm] = startValue.split(':').map(Number);
-    const end = new Date();
-    end.setHours(sh + duration, sm);
-    endValue = `${pad2(end.getHours())}:${pad2(end.getMinutes())}`;
-  }
-
-  const overlay = document.createElement('div');
-  overlay.className = 'manage-overlay';
-  overlay.innerHTML = `
-    <div class="manage-dialog">
-      <div class="manage-header">
-        <div class="role-chip">${st.role === 'nurse' ? 'RN' : 'TECH'}</div>
-        <h3>${st.name || labelFromId(st.id)}</h3>
-        <div class="staff-type">${st.type}</div>
-      </div>
-      <div class="manage-grid">
-        <div class="manage-col">
-          <label>Start <input id="mg-start" type="time" value="${startValue}" disabled></label>
-          <label>End <input id="mg-end" type="time" required value="${endValue}"></label>
-          <label>RF <input id="mg-rf" type="number" value="${st.rf ?? ''}"></label>
-        </div>
-        <div class="manage-col">
-          <label class="mod"><span class="icon">🎓</span><input id="mg-student" placeholder="Student" value="${typeof slot.student === 'string' ? slot.student : ''}"></label>
-          <label class="mod"><span class="icon">☕</span><input type="checkbox" id="mg-break" ${slot.break?.active ? 'checked' : ''}/> Break</label>
-          <label class="mod"><span class="icon">🔥</span><input type="checkbox" id="mg-high" ${slot.highAcuityUntil && slot.highAcuityUntil > Date.now() ? 'checked' : ''}/> High acuity</label>
-          <label class="mod">Zone <select id="mg-zone">
-            ${(cfg.zones || [])
-              .map((z: ZoneDef) => `<option value="${z.name}"${z.name === zone ? ' selected' : ''}>${z.name}</option>`)
-              .join('')}
-          </select></label>
-        </div>
-      </div>
-      <div class="dialog-actions">
-        <button id="mg-save" class="btn">Save</button>
-        <button id="mg-cancel" class="btn">Cancel</button>
-      </div>
-    </div>
-  `;
-  document.body.appendChild(overlay);
-
-  overlay.querySelector('#mg-cancel')!.addEventListener('click', () => overlay.remove());
-
-  overlay.querySelector('#mg-save')!.addEventListener('click', async () => {
-    const rfVal = (overlay.querySelector('#mg-rf') as HTMLInputElement).value.trim();
-    st.rf = rfVal ? Number(rfVal) : undefined;
-
-    const studVal = (overlay.querySelector('#mg-student') as HTMLInputElement).value.trim();
-    slot.student = studVal ? studVal : undefined;
-
-    const breakChecked = (overlay.querySelector('#mg-break') as HTMLInputElement).checked;
-    if (breakChecked && !slot.break?.active) startBreak(slot, {});
-    if (!breakChecked && slot.break?.active) endBreak(slot);
-
-    const highChecked = (overlay.querySelector('#mg-high') as HTMLInputElement).checked;
-    slot.highAcuityUntil = highChecked ? Date.now() + 2 * 60 * 60 * 1000 : undefined;
-
-    const endVal = (overlay.querySelector('#mg-end') as HTMLInputElement).value;
-    slot.endTimeOverrideHHMM = endVal;
-
-    const zoneSel = overlay.querySelector('#mg-zone') as HTMLSelectElement;
-    if (zoneSel.value !== zone) {
-      const moved = moveSlot(board, { zone, index }, { zone: zoneSel.value });
-      if (moved) showBanner('Previous assignment cleared');
-    }
-
-    await saveStaff(staffList);
-    save();
-    overlay.remove();
-    rerender();
-  });
-}
-
